Make the IAM policy creation optional
In cases where projects wish to avoid role chaining, an IAM policy for
cross-account role assumption can be attached directly to an IAM user or
group. In that scenario, it is not necessary for this module to attach
an additional cross-account IAM policy to the role that this module
creates.

https://www.pivotaltracker.com/story/show/170449919